# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Simpsonm09 Fixture Rust API",
    "description": "Item CRUD service for the simpsonm09 repository fixture",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Lists every item.",
        "operationId": "list_items",
        "responses": {
          "200": {
            "description": "List every item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Creates an item.",
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Create an item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Gets one item by id.",
        "operationId": "get_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Item identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Get one item by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Replaces an item.",
        "operationId": "update_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Item identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replace an item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Deletes an item.",
        "operationId": "delete_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Item identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Item deleted"
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ItemRequest": {
        "type": "object",
        "description": "Payload used to create or replace an item. `id` is assigned by the store.",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Item description.",
            "example": "A small widget",
            "maxLength": 2000
          },
          "name": {
            "type": "string",
            "description": "Item name.",
            "example": "Widget",
            "maxLength": 200
          }
        }
      },
      "ItemResponse": {
        "type": "object",
        "description": "An item returned by the API.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Item description.",
            "example": "A small widget"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-assigned identifier.",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Item name.",
            "example": "Widget"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "description": "RFC 7807 problem detail returned for item errors.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Explanation specific to this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary."
          },
          "type": {
            "type": "string",
            "description": "Problem type URI."
          }
        }
      }
    }
  }
}